third_party/WebKit/LayoutTests/fast/css/add-remove-stylesheets-minimal-recalc-style-expected.txt

PASS internals.updateStyleAndReturnAffectedElementCount() is 0
PASS internals.updateStyleAndReturnAffectedElementCount() is 0
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS internals.updateStyleAndReturnAffectedElementCount() is 31
PASS internals.updateStyleAndReturnAffectedElementCount() is 3
PASS internals.updateStyleAndReturnAffectedElementCount() is 3
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS internals.updateStyleAndReturnAffectedElementCount() is 3
PASS successfullyParsed is true

TEST COMPLETE
foo
bar
baz
